+++ simply-synonyms/popup/popup.js
@@ -1,6 +1,8 @@
 let onlyEditableText = document.getElementById('only_editable_text_switch')
 let disablePopup = document.getElementById('disable_switch')
 
+
+/* SETTINGS AND OPTIONS */
 function settingsChanged() {
   chrome.storage.local.set({
     option_popupDisabled: disablePopup.checked,
@@ -15,7 +17,6 @@ chrome.storage.local.get(['option_popupDisabled', 'option_onlyEditableText'], (r
   onlyEditableText.checked = result.option_onlyEditableText
 })
 
-
 onlyEditableText.addEventListener('click', settingsChanged)
 disablePopup.addEventListener('click', settingsChanged)
 
@@ -23,6 +24,48 @@ document.getElementById('version-text').innerText = `V${chrome.runtime.getManife
 
 if (!('update_url' in chrome.runtime.getManifest())) document.getElementById('dev-badge').style.display = 'block'
 
+/* AUTHENTICATION */
+const googleSigninButton = document.getElementById('google-signin')
+const signinDiv = document.getElementById('signin-div')
+const signoutButton = document.getElementById('signout')
+const userWelcome = document.getElementById('user-welcome')
+
+const welcomeMessages = [
+  'Are you ready to start writing?',
+  'Are you ready to find some great words?',
+  'Are you looking forward to using Simply Synonyms today?',
+  'How has your day been so far?',
+  'Did you know you can see a list of upcoming features at the bottom of this page?',
+  'Did you know that Simply Synonyms is open source, meaning anyone who knows how to code can contribute to it?',
+  'What great words are you going to find today?',
+  'Thanks for using Simply Synonyms!',
+  'What a beautiful day.',
+  ''
+]
+
+firebase.auth().onAuthStateChanged((user) => {
+  if (user) {
+    userWelcome.innerText = `Hello, ${user.displayName}. ${welcomeMessages[Math.floor(Math.random() * welcomeMessages.length)]}`
+    signoutButton.classList.remove('hidden')
+    signinDiv.classList.add('hidden')
+  } else {
+    signoutButton.classList.add('hidden')
+    signinDiv.classList.remove('hidden')
+    userWelcome.innerText = ''
+  }
+  googleSigninButton.disabled = false
+});
+
+googleSigninButton.addEventListener('click', (e) => {
+    googleSigninButton.disabled = true
+    chrome.runtime.sendMessage({ action: 'getAuthToken', interactive: true})
+})
+
+signoutButton.addEventListener('click', (e) => {
+  chrome.runtime.sendMessage({ action: 'signOut' })
+})
+
+/* QUICK SEARCH */
 document.getElementById('quicksearch').addEventListener('input', (e) => {
   const quicksearchPrompt = document.getElementById('quicksearch-prompt')
   if (e.target.value.length !== 0) {